add more exception handling and health check

+++ src/MinimalUserAPI.Infrastructure/Users/UserRepository.cs
@@ -14,6 +14,10 @@ public class UserRepository : IUserRepository
 
     public async ValueTask<long> DeleteUser(int userId)
     {
+        if (userId <= 0)
+        {
+            throw new ArgumentException($"{nameof(userId)} must be greater than '0'");
+        }
         var deleteResult = await dbContext.Users.DeleteOneAsync(s => s.Id == userId);
         return deleteResult.DeletedCount;
     }
@@ -27,6 +31,7 @@ public class UserRepository : IUserRepository
 
     public async ValueTask<User> InsertUser(User user)
     {
+        ArgumentNullException.ThrowIfNull(user, nameof(user)); 
         await dbContext.Users.InsertOneAsync(user);
         return user;
 
@@ -34,6 +39,11 @@ public class UserRepository : IUserRepository
 
     public async ValueTask<User> UpdateUser(int userId,User user)
     {
+        if (userId <= 0)
+        {
+            throw new ArgumentException($"{nameof(userId)} must be greater than '0'");
+        }
+        ArgumentNullException.ThrowIfNull(user, nameof(user));
         var replaceResult = await dbContext.Users.ReplaceOneAsync(s => s.Id == userId, user);
         if (replaceResult.ModifiedCount == 0)
         {
